requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.3
Jinja2==2.6
SQLAlchemy==0.7.10
Tempita==0.5.1
WTForms==1.0.3
Werkzeug==0.8.3
argparse==1.2.1
decorator==3.4.0
distribute==0.6.24
psycopg2==2.4.6
sqlalchemy-migrate==0.7.2
sqlsoup==0.9.0
wsgiref==0.1.2
